+++ web/apps/nextjs__raw_krabby_chat_api_integrations/app/(routes)/(chat)/[chatId]/(call-routes)/video/page.tsx
@@ -12,7 +12,7 @@ import {
   CameraOffIcon,
   EndCallIcon,
   SettingsIcon,
-} from '../../../../../Icons';
+} from '../../../../../icons';
 import MiniSoundStreamVisualizer from './components/MiniSoundStreamVisualizer';
 import RTCDebugOverlay from './components/RTCDebugOverlay';
